# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Documentation",
        "description": "API Swagger description for HelloCSE Application test",
        "contact": {
            "email": "contact@obbydev.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new administrator",
                "operationId": "eb3623f9bcd09a1dacbefaac07aed1bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully!"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/Administrator"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "token_value"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate an administrator",
                "operationId": "f77777235d676b651c9465807d8f9485",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User login successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User login successfully!"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/Administrator"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "token_value"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout the authenticated administrator",
                "operationId": "c4b3c6f1776909e546a421283c044ac3",
                "responses": {
                    "200": {
                        "description": "User logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User logged out successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Administrators"
                ],
                "summary": "Get administrator information",
                "description": "Returns administrator data",
                "operationId": "facf643270f6f524d7a721a4018aa7fb",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Administrator"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Retrieve a list of active profiles",
                "operationId": "d7a57f4d7de18fb0f92501172472c9b6",
                "responses": {
                    "200": {
                        "description": "A list of active profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Profil"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Create a new profile",
                "operationId": "9fea6f2adcb69614c521a7733358382d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfilRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profil"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Retrieve a specific profile",
                "operationId": "40a5104e23b52137e78c3d3ac48b88a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the profile to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profil"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update an existing profile",
                "operationId": "b166772cec9714067040e77cf14beeba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the profile to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfilRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profil"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    },
                    "500": {
                        "description": "Failed to update profile"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Delete a profile",
                "operationId": "ce06860a51ab016ef66e753544fe4563",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the profile to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profil successfully deleted!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    },
                    "500": {
                        "description": "Failed to delete profile"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Administrator": {
                "title": "Administrator",
                "description": "Administrator model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the administrator",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the administrator",
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of the administrator",
                        "type": "string",
                        "format": "password",
                        "example": "hashedpassword123"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email address of the administrator",
                        "type": "string",
                        "example": "jane.doe@example.com"
                    },
                    "profiles": {
                        "title": "Profiles",
                        "description": "List of profiles associated with the administrator",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Profil"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Administrator"
                }
            },
            "Profil": {
                "title": "Profil",
                "description": "Profil model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the profil",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "firstname": {
                        "title": "First Name",
                        "description": "First name of the profil",
                        "type": "string",
                        "example": "John"
                    },
                    "lastname": {
                        "title": "Last Name",
                        "description": "Last name of the profil",
                        "type": "string",
                        "example": "Doe"
                    },
                    "picture": {
                        "title": "Picture",
                        "description": "URL of the profil picture",
                        "type": "string",
                        "example": "http://example.com/images/profile.jpg"
                    },
                    "status": {
                        "title": "Status",
                        "description": "Status of the profil",
                        "type": "string",
                        "example": "active"
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "ID of the associated user",
                        "type": "integer",
                        "format": "int64",
                        "example": 10
                    },
                    "administrator": {
                        "$ref": "#/components/schemas/Administrator"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Profil"
                }
            },
            "ProfilRequest": {
                "title": "ProfilRequest",
                "description": "Request body data for creating or updating a Profil",
                "required": [
                    "firstname",
                    "lastname"
                ],
                "properties": {
                    "firstname": {
                        "title": "First Name",
                        "description": "The first name of the profile",
                        "type": "string",
                        "example": "John"
                    },
                    "lastname": {
                        "title": "Last Name",
                        "description": "The last name of the profile",
                        "type": "string",
                        "example": "Doe"
                    },
                    "status": {
                        "title": "Status",
                        "description": "The status of the profile",
                        "type": "string",
                        "example": "active"
                    },
                    "picture": {
                        "title": "Picture",
                        "description": "Profile picture file",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Administrators",
            "description": "Administrators"
        },
        {
            "name": "Profiles",
            "description": "Profiles"
        }
    ]
}